Extract isaOutput scss
